pyproject.toml
[tool.poetry]
name = "render-rabbitmq"
version = "0.1.0"
description = ""
authors = ["Stharley Maxwell <stharleymax@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pika = "^1.3.2"
python-decouple = "^3.8"
black = "^23.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
